+++ lib/Braintree/Transaction.php
@@ -176,14 +176,15 @@ class Transaction extends Base
     const RECURRING     = 'recurring';
 
     // Gateway Rejection Reason
-    const AVS            = 'avs';
-    const AVS_AND_CVV    = 'avs_and_cvv';
-    const CVV            = 'cvv';
-    const DUPLICATE      = 'duplicate';
-    const FRAUD          = 'fraud';
-    const RISK_THRESHOLD = 'risk_threshold';
-    const THREE_D_SECURE = 'three_d_secure';
-    const TOKEN_ISSUANCE = 'token_issuance';
+    const AVS                    = 'avs';
+    const AVS_AND_CVV            = 'avs_and_cvv';
+    const CVV                    = 'cvv';
+    const DUPLICATE              = 'duplicate';
+    const EXCESSIVE_RETRY        = 'excessive_retry';
+    const FRAUD                  = 'fraud';
+    const RISK_THRESHOLD         = 'risk_threshold';
+    const THREE_D_SECURE         = 'three_d_secure';
+    const TOKEN_ISSUANCE         = 'token_issuance';
     const APPLICATION_INCOMPLETE = 'application_incomplete';
 
     // Industry Types