+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
-## 6.11.0
+## 6.12.0
 * Address PHP 8.2 deprecation warnings for dynamic properties usage.
+
+## 6.11.0 
+* Add `preDisputeProgram` to `Dispute` and `DisputeSearch`
+* Add `AUTO_ACCEPTED` Dispute Status
+* Add `DISPUTE_AUTO_ACCEPTED` WebhookNotification
+* Add `EXCESSIVE_RETRY` to `GatewayRejectionReason`
 
 ## 6.10.0
 * Add `SEPADirectDebitAccount` payment method